// pdlearning/src/web-apps/src/libraries/domain-api/src/calendar/enums/event-source.ts
export enum EventSource {
  SelfCreated = 'SelfCreated',
  LNA = 'LNA',
  CourseClassRun = 'CourseClassRun',
  StandaloneForm = 'StandaloneForm',
  ExternalPDO = 'ExternalPDO',
  Webinar = 'CommunityWebinar',
  Community = 'CommunityRegular',
  CourseSession = 'CourseSession',
  CourseAssignment = 'CourseAssignment'
}